Sign in if Authenticated
This adds a way to auto login if the user has a valid token. The only
downside is that this will only work for the first 2 hours before having
to login again after closing the app. We may want to try a refresh
before directing them to the login at a later point.

Also fix a bug where the expires_in was not being returned. This
resulted in all requests getting a new access token.